backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic[email]
python-dotenv==1.0.0
httpx==0.26.0
aiohttp==3.9.0
beautifulsoup4==4.12.3
selenium==4.16.0
webdriver-manager==4.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
bcrypt==4.1.2
pyjwt==2.8.0